Move Protocol to backends
To avoid circular imports while still typing both Transport and MachineState
